SCRATCH: HTML values still not read with /q
I'm not sure why, but input values aren't read from HTML pages
when htmlInstaller.exe is launched with /q.

The current test uses a value of 3 in the configuration.xml and
a value of 4 in the HTML. I was expecting the 4 to overwrite the 3,
but that only seems to happen if I run the linked installer after
the test has built it.